Improved how we ask in the queue for new recordings.
If we have no new message in the queue, then we will wait for 5 seconds
and then check again.
But if we find one, we will immediately ask for another one.
This way, if we get a bunch of messages in the queue, they are all
processed relatively fast.
And if we have none, then we don't ask for a new message a thousand
times.